requirements.txt
together
pandas
numpy
matplotlib 
seaborn 
python-docx 
PyPDF2 
openpyxl 
streamlit
dotenv
